requirements.txt
asgiref==3.5.0
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
platformdirs==2.5.1
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.1
whitenoise==6.0.0
